+++ src/hittable_list.cuh
@@ -10,14 +10,14 @@ class hittable_list : public hittable {
     __device__ hittable_list() {}
     __device__ hittable_list(hittable **l, int n) : list(l), list_size(n) {}
     __device__ virtual bool hit(
-        const ray& r, double tmin, double tmax, hit_record& rec) const override;
+        const ray& r, float tmin, float tmax, hit_record& rec) const override;
 
   public:
     hittable **list;
     int list_size;
 };
 
-__device__ bool hittable_list::hit(const ray&r, double t_min, double t_max, hit_record& rec) const {
+__device__ bool hittable_list::hit(const ray&r, float t_min, float t_max, hit_record& rec) const {
   hit_record temp_rec;
   bool hit_anything = false;
   auto closest_so_far = t_max;
