+++ src/sphere.cuh
@@ -6,16 +6,16 @@
 class sphere : public hittable {
   public:
     __device__ sphere() {}
-    __device__ sphere(point3 cen, double r) : center(cen), radius(r) {};
+    __device__ sphere(point3 cen, float r) : center(cen), radius(r) {};
     __device__ virtual bool hit(
-        const ray& r, double tmin, double tmax, hit_record& rec) const override;
+        const ray& r, float tmin, float tmax, hit_record& rec) const override;
 
   public:
     point3 center;
-    double radius;
+    float radius;
 };
 
-__device__ bool sphere::hit(const ray& r, double t_min, double t_max, hit_record& rec) const {
+__device__ bool sphere::hit(const ray& r, float t_min, float t_max, hit_record& rec) const {
   vec3 oc = r.origin() - center;
   auto a = r.direction().length_squared();
   auto half_b = dot(oc, r.direction());
